+++ packages/angular-query-persist-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tanstack/angular-query-persist-client
 
+## 5.101.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/angular-query-experimental@5.101.4
+  - @tanstack/query-persist-client-core@5.101.4
+
 ## 5.101.3
 
 ### Patch Changes