+++ packages/query-devtools/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @tanstack/query-devtools
 
+## 5.101.4
+
 ## 5.101.3
 
 ## 5.101.2